// src/http/generated/models/AdvanceFilterRequestModel.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MonoWatch BackBone
 * MonoWatch Security Asset Management
 * OpenAPI spec version: 1.0.0
 */
import type { AdvanceFilterRequestModelSearch } from './AdvanceFilterRequestModelSearch';
import type { AdvanceFilterColumnsRequest } from './AdvanceFilterColumnsRequest';
import type { AdvanceFilterRequestModelGroupBy } from './AdvanceFilterRequestModelGroupBy';
import type { AdvanceFilterRequestModelConditions } from './AdvanceFilterRequestModelConditions';
import type { AdvanceFilterRequestModelStartDate } from './AdvanceFilterRequestModelStartDate';
import type { AdvanceFilterRequestModelEndDate } from './AdvanceFilterRequestModelEndDate';

export interface AdvanceFilterRequestModel {
  search?: AdvanceFilterRequestModelSearch;
  columns: AdvanceFilterColumnsRequest[];
  group_by?: AdvanceFilterRequestModelGroupBy;
  conditions?: AdvanceFilterRequestModelConditions;
  start_date?: AdvanceFilterRequestModelStartDate;
  end_date?: AdvanceFilterRequestModelEndDate;
  page: number;
  limit: number;
}