// src/http/generated/models/AdvancedFilterColumnInformation.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MonoWatch BackBone
 * MonoWatch Security Asset Management
 * OpenAPI spec version: 1.0.0
 */
import type { AdvancedFilterColumnInformationObjectType } from './AdvancedFilterColumnInformationObjectType';
import type { AdvancedFilterColumnInformationOptions } from './AdvancedFilterColumnInformationOptions';

export interface AdvancedFilterColumnInformation {
  name: string;
  display_name: string;
  type: string;
  is_default: boolean;
  object_type?: AdvancedFilterColumnInformationObjectType;
  options?: AdvancedFilterColumnInformationOptions;
}